feat: 채팅방 사진 업로드 UI 수정
- 2-5개일 경우: 동일 width값이되, 그리드 계산해서 보여주기
- 1개일 경우: 동일 width값이되, 비율 유지
- 전송 전 키보드 상단에 전송할 이미지 미리 보기
